+++ src/PGMDozer.hpp
@@ -11,6 +11,6 @@ using namespace Eigen;
 
 // typedef Matrix< float , 1, Dynamic > Eigen::RowVectorXf
 // typedef Matrix< double , 1, Dynamic > Eigen::RowVectorXd
-void get_image(const char* image_route, unsigned int ancho, unsigned int alto, RowVectorXf& imagen);
-void save_image(const char* image_route, unsigned int ancho, unsigned int alto, RowVectorXf& imagen);
+void get_image(const char* image_route, unsigned int ancho, unsigned int alto, RowVectorXd& imagen);
+void save_image(const char* image_route, unsigned int ancho, unsigned int alto, RowVectorXd& imagen);
 void show_help();
